Move batchWriteRetry() to utils for re-use
