+++ src/putAll.spec.js
@@ -8,7 +8,7 @@ const mockClient = (method = 'batchWrite') => {
   }
 }
 
-const { appendPutAll, batchWriteRetry } = require('./putAll')
+const { appendPutAll } = require('./putAll')
 
 beforeEach(() => {
   jest.clearAllMocks()
@@ -22,43 +22,6 @@ describe('appendPutAll()', () => {
 
     expect(client).toHaveProperty('putAll')
     expect(client.putAll).toBeInstanceOf(Function)
-  })
-})
-
-describe('batchWriteRetry()', () => {
-  it('returns a Promise', async () => {
-    const client = mockClient()
-
-    expect(batchWriteRetry(client)).toBeInstanceOf(Promise)
-  })
-
-  it('forwards params to the client', async () => {
-    const client = mockClient()
-
-    const TableName = 'Area51'
-    const params = { RequestItems: { [TableName]: [] } }
-    await batchWriteRetry(client, params)
-
-    expect(client.mockReference).toHaveBeenCalledWith(params)
-  })
-
-  it('retries unprocessed items', async () => {
-    const client = mockClient()
-
-    const TableName = 'Area51'
-    const Item = {
-      id: 'foo',
-      name: 'Matthew'
-    }
-    const RequestItems = { [TableName]: [{ PutRequest: { Item } }] }
-
-    // Make sure it fails once.
-    client.mockReference.mockResolvedValueOnce({ UnprocessedItems: RequestItems })
-
-    const params = { RequestItems }
-    await batchWriteRetry(client, params)
-
-    expect(client.mockReference).toHaveBeenCalledTimes(2)
   })
 })
